Extract errors to separate module
