--- app/services/google/search_service_error.rb
@@ -1,13 +1,0 @@
-module Google
-  class SearchServiceError < StandardError
-    attr_reader :original_error
-
-    def initialize(url, error = nil)
-      @original_error = error
-      msg = "Failed to perform search request with URL: #{url}"
-      super(msg)
-    end
-
-    def to_s = "#{super}\nOriginal error: #{@original_error}"
-  end
-end
